Update program 04.py

+++ ProgramBatch2/program 04.py
@@ -1,3 +1,3 @@
-num1 = float(input("Enter first number: "))
-num2 = float(input("Enter second number: "))
+num1 = int(input("Enter first number: "))
+num2 = int(input("Enter second number: "))
 print("Quotient:", num1 // num2)
